+++ src/main/kotlin/com/alexhalcazar/mydrinksapp/model/DrinkResponse.kt
@@ -1,0 +1,9 @@
+package com.alexhalcazar.mydrinksapp.model
+
+import com.alexhalcazar.mydrinksapp.model.Drink
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class DrinkResponse(
+    val drinks: List<Drink>?
+)
